test: address adversarial review findings (issue #360)
