+++ src/Profiles/DataGenerator.cs
@@ -17,11 +17,33 @@ internal class DataGenerator
     private readonly DateTime now;
     private int documentIndex;
 
-    public DataGenerator(ColumnProfile profile, int? seed = null, DateTime? now = null, int? custodianCountOverride = null)
+    public DataGenerator(
+        ColumnProfile profile,
+        int? seed = null,
+        DateTime? now = null,
+        int? custodianCountOverride = null,
+        string? dateFormatOverride = null,
+        int? emptyPercentageOverride = null)
     {
-        this.profile = custodianCountOverride.HasValue
+        var tempProfile = custodianCountOverride.HasValue
             ? ApplyCustodianCountOverride(profile, custodianCountOverride.Value)
             : profile;
+
+        this.profile = tempProfile;
+        if (!string.IsNullOrEmpty(dateFormatOverride))
+        {
+            this.profile.Settings.DateFormat = dateFormatOverride;
+        }
+
+        if (emptyPercentageOverride.HasValue)
+        {
+            this.profile.Settings.EmptyValuePercentage = emptyPercentageOverride.Value;
+            foreach (var col in this.profile.Columns)
+            {
+                col.EmptyPercentage = emptyPercentageOverride.Value;
+            }
+        }
+
 #pragma warning disable S2245
         this.random = seed.HasValue ? new Random(seed.Value) : Random.Shared;
 #pragma warning restore S2245
